# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Char Par game platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface GuestLoginInput {
  username?: string;
}

export interface GoogleLoginInput {
  credential: string;
}

export interface User {
  id: string;
  username: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatar?: string | null;
  isGuest: boolean;
  eloRating: number;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserStats {
  totalGames: number;
  wins: number;
  losses: number;
  winRate: number;
  currentStreak: number;
  bestStreak: number;
  rankedGames: number;
  rankedWins: number;
}

export interface UserProfile {
  id: string;
  username: string;
  /** @nullable */
  avatar?: string | null;
  eloRating: number;
  isGuest: boolean;
  stats: UserStats;
  createdAt: string;
}

export type MatchRecordResult = typeof MatchRecordResult[keyof typeof MatchRecordResult];


export const MatchRecordResult = {
  win: 'win',
  loss: 'loss',
} as const;

export type MatchRecordMode = typeof MatchRecordMode[keyof typeof MatchRecordMode];


export const MatchRecordMode = {
  ranked: 'ranked',
  casual: 'casual',
  ai: 'ai',
  local: 'local',
} as const;

export interface MatchRecord {
  id: string;
  opponentId: string;
  opponentUsername: string;
  /** @nullable */
  opponentAvatar?: string | null;
  result: MatchRecordResult;
  mode: MatchRecordMode;
  /** @nullable */
  eloDelta?: number | null;
  /** @nullable */
  duration?: number | null;
  playedAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  username: string;
  /** @nullable */
  avatar?: string | null;
  eloRating: number;
  wins: number;
  losses: number;
  winRate: number;
  currentStreak: number;
}

export interface LeaderboardSummary {
  totalPlayers: number;
  totalGamesPlayed: number;
  topElo: number;
  averageElo: number;
}

export type GameMode = typeof GameMode[keyof typeof GameMode];


export const GameMode = {
  ranked: 'ranked',
  casual: 'casual',
  ai: 'ai',
  local: 'local',
} as const;

export type GameStatus = typeof GameStatus[keyof typeof GameStatus];


export const GameStatus = {
  waiting: 'waiting',
  active: 'active',
  completed: 'completed',
  abandoned: 'abandoned',
} as const;

export interface Game {
  id: string;
  player1Id: string;
  player1Username: string;
  /** @nullable */
  player2Id?: string | null;
  /** @nullable */
  player2Username?: string | null;
  mode: GameMode;
  status: GameStatus;
  /** @nullable */
  winnerId?: string | null;
  /** @nullable */
  duration?: number | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export type GetLeaderboardParams = {
mode?: GetLeaderboardMode;
limit?: number;
};

export type GetLeaderboardMode = typeof GetLeaderboardMode[keyof typeof GetLeaderboardMode];


export const GetLeaderboardMode = {
  ranked: 'ranked',
  casual: 'casual',
} as const;

export type ListGamesParams = {
limit?: number;
};