# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Char Par game platform API
 * OpenAPI spec version: 0.1.0
 */
import type { GameMode } from './gameMode';
import type { GameStatus } from './gameStatus';

export interface Game {
  id: string;
  player1Id: string;
  player1Username: string;
  /** @nullable */
  player2Id?: string | null;
  /** @nullable */
  player2Username?: string | null;
  mode: GameMode;
  status: GameStatus;
  /** @nullable */
  winnerId?: string | null;
  /** @nullable */
  duration?: number | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}