# lib/api-zod/src/generated/types/leaderboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Char Par game platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardSummary {
  totalPlayers: number;
  totalGamesPlayed: number;
  topElo: number;
  averageElo: number;
}