# lib/api-zod/src/generated/types/matchRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Char Par game platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchRecordMode } from './matchRecordMode';
import type { MatchRecordResult } from './matchRecordResult';

export interface MatchRecord {
  id: string;
  opponentId: string;
  opponentUsername: string;
  /** @nullable */
  opponentAvatar?: string | null;
  result: MatchRecordResult;
  mode: MatchRecordMode;
  /** @nullable */
  eloDelta?: number | null;
  /** @nullable */
  duration?: number | null;
  playedAt: string;
}