# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Char Par game platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  username: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatar?: string | null;
  isGuest: boolean;
  eloRating: number;
  createdAt: string;
}