# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Char Par game platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserStats } from './userStats';

export interface UserProfile {
  id: string;
  username: string;
  /** @nullable */
  avatar?: string | null;
  eloRating: number;
  isGuest: boolean;
  stats: UserStats;
  createdAt: string;
}